move platform stuff to update and tweak updater switch
