+++ app/__main__.py
@@ -2,11 +2,11 @@ import asyncio
 import os.path
 import sys
 
-from PySide6.QtCore import QTranslator, QLockFile
+from PySide6.QtCore import QTranslator, QLocale, QLockFile
 from qasync import QApplication, run
 
 from app.builtin.locale import detect_system_ui_language
-from app.builtin.utils import get_updater, running_in_bundle, init_app
+from app.builtin.utils import get_updater, init_app
 from app.builtin.paths import AppPaths
 from app.main_window import MainWindow
 
@@ -32,7 +32,7 @@ def main(enable_updater: bool = True):
     # and do update logic
     updater = get_updater()
     # self-updating is not available on macOS
-    updater.is_enable = False if running_in_bundle else enable_updater
+    # updater.is_enable = False if running_in_bundle else enable_updater
 
     # override updater config
     if os.getenv("DEBUG", "0") == 1 and os.path.exists("updater.json"):
