+++ app/builtin/gitlab_updater.py
@@ -1,5 +1,4 @@
 import os
-import platform
 from urllib.parse import urlparse
 
 from glom import glom
@@ -7,9 +6,8 @@ from httpx import AsyncClient
 
 from singleton_decorator import singleton
 
-from app.builtin.update import Updater, Version
+from app.builtin.update import Updater, Version, get_arch, get_sysname
 from app.builtin.paths import AppPaths
-from app.builtin.utils import get_arch, get_sysname
 
 
 @singleton
