+++ app/builtin/utils.py
@@ -3,7 +3,6 @@ from app.builtin.gitlab_updater import GitlabUpdater
 import app.builtin.config as cfg
 
 import sys
-import platform
 from pathlib import Path
 
 from qdarktheme import enable_hi_dpi
@@ -32,30 +31,6 @@ def running_in_bundle() -> bool:
 
     exe_path = Path(sys.executable).resolve()
     return ".app/Contents/MacOS" in str(exe_path)
-
-
-def get_sysname() -> str:
-    sysname = platform.system().lower()
-    if sysname == "windows":
-        sysname = "windows"
-    elif sysname == "darwin":
-        sysname = "macos"
-    elif sysname == "linux":
-        sysname = "linux"
-    else:
-        raise RuntimeError(f"Unknown system: {sysname}")
-    return sysname
-
-
-def get_arch() -> str:
-    arch = platform.machine().lower()
-    if arch in ["x86_64", "amd64"]:
-        arch = "x64"
-    elif arch in ["aarch64", "arm64"]:
-        arch = "arm64"
-    else:
-        raise RuntimeError(f"Unknown architecture: {arch}")
-    return arch
 
 
 def init_app():
